+++ CCP/settings.py
@@ -123,3 +123,8 @@ AUTH_USER_MODEL = 'user_info.User'
 # https://docs.djangoproject.com/en/1.11/howto/static-files/
 
 STATIC_URL = '/static/'
+
+STATICFILES_DIRS = [
+    os.path.join(BASE_DIR, "static"),
+    '/var/www/static/',
+]
