--- main_site/templates/base.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<head>
-    <meta charset="UTF-8">
-    <title>信息与电子学院党建信息平台</title>
-</head>
-<body>
-{% block content %}
-{% endblock %}
-</body>
-</html>
