--- main_site/templates/index.html
@@ -1,3 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-{% endblock %}
