+++ main_site/templates/main_site/base.html
@@ -1,0 +1,67 @@
+<!DOCTYPE html>
+<head>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
+    <meta charset="UTF-8">
+    <title>信息与电子学院党建信息平台</title>
+</head>
+<body>
+<style>
+    @media (min-width: 1300px) {
+        .container {
+            max-width: 1300px;
+        }
+    }
+</style>
+
+<div class="container mt-3">
+    {% load static %}
+    <img src="{% static "logo.png" %}"/>
+</div>
+<div class="container mt-3">
+    <hr style="margin:0 !important; border:none;border-top:1px solid #cc1c25;"/>
+
+    <nav class="navbar navbar-expand-lg navbar-light bg-white">
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarTogglerDemo03" aria-controls="navbarTogglerDemo03" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarTogglerDemo03">
+            <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
+                <li class="nav-item">
+                    <a class="nav-link" href="#" style="color: #8a1010;font-weight:bold;">首页</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#" style="color: #8a1010;font-weight:bold;">我的信息</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#" style="color: #8a1010">入党流程</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#" style="color: #8a1010">志愿时长</a>
+                </li>
+            </ul>
+            <div class="form-inline my-2 my-lg-0">
+                {% if user.is_authenticated %}
+                    <li class="nav-item">
+                        <a class="nav-link" href="settings/{{ user.username }}">Welcome {{ user.username }} !</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link" href="/logout">注销</a>
+                    </li>
+                {% else %}
+                    <a class="nav-link" href="{% url 'login' %}" style="color:#333;">登录</a>
+                    <a class="nav-link" href="{% url 'register' %}" style="color:#333;">注册</a>
+                {% endif %}
+            </div>
+        </div>
+    </nav>
+    <hr style="margin:0 !important; border:none;border-top:4px solid #cc1c25;"/>
+</div>
+
+
+{% block content %}
+{% endblock %}
+</body>
+</html>
