+++ main_site/templates/main_site/index.html
@@ -1,0 +1,10 @@
+{% extends "main_site/base.html" %}
+{% block content %}
+    <div class="container">
+        <div style="overflow:hidden;width:1300px;height:2125px;">
+            <div style="margin-left: -55px;margin-top: -230px;margin-right: -30px;">
+                <iframe allowtransparency="true" class="ui-layout-center" scrolling="no" src="http://www.12371.cn/" id="mainframe" name="mainframe" style="width:100%;height:5000px;"></iframe>
+            </div>
+        </div>
+    </div>
+{% endblock %}
