+++ main_site/templates/main_site/register.html
@@ -1,0 +1,6 @@
+{% extends "main_site/base.html" %}
+{% block content %}
+    <div class="container">
+
+    </div>
+{% endblock %}
